refactor: extract CSV parsing into hook
